% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{highs_available_solver_options}
\alias{highs_available_solver_options}
\title{Available Solver Options}
\usage{
highs_available_solver_options()
}
\value{
A \code{data.frame} containing the available solver options.
}
\description{
Reference for the available solver options.
}
\examples{
highs_available_solver_options()
}
