% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highs.R
\name{highs_control}
\alias{highs_control}
\title{Highs Control}
\usage{
highs_control(threads = 1L, time_limit = Inf, log_to_console = FALSE, ...)
}
\arguments{
\item{threads}{an integer giving the number of threads to be used.}

\item{time_limit}{a double giving the time limit.}

\item{log_to_console}{a logical giving if the output should be shown in
the console.}

\item{...}{other arguments supported by the \code{HiGHS} solver.}
}
\description{
Highs Control
}
\examples{
control <- highs_control()
}
